The change avoids unnecessary pushes.
We are running into serious problems with the following scenario: When a user logs in the first time, we create a user specific couch db and insert some design documents to that db. this is realized via a separate node process with nano.

In this case, the couch-daemon (we use couchmagick) gets a massive bulk of events (creation of new db, updates for several design documents, etc.). I don't know why, but then the new db is not watched by couch-daemon, if if the dbs name is covered by the whitelist (/.*/ on our case).

Applying the above patch avoids the problem.